TApiResponse修正、article/createで動作確認

+++ src/pages/api/user_info/delete.ts
@@ -1,16 +1,15 @@
-import { db } from "../../../lib/db";
-import { NextApiRequest, NextApiResponse } from "next";
-import { server, localhost } from "../../../lib/loadUrl";
-import { TApiResponse } from "../../../lib/apiTypes";
-import { T_user_id } from "../../../app/Store/Interface";
-import { apiWrapPost } from "../../../lib/apiWrap";
-
+import { db } from '../../../lib/db';
+import { NextApiRequest, NextApiResponse } from 'next';
+import { server, localhost } from '../../../lib/loadUrl';
+import { TApiResponse } from '../../../lib/apiWrap';
+import { T_user_id } from '../../../app/Store/Interface';
+import { apiWrapPost } from '../../../lib/apiWrap';
 
 // サーバーサイドとフロントサイド考えずに使えるようにラップする
 export const apiUserInfoDelete = async (
   params: T_user_info_delete
 ): Promise<TApiResponse<T_user_info_delete_return>> => {
-  return apiWrapPost("user_info/delete", params);
+  return apiWrapPost('user_info/delete', params);
 };
 
 export type T_user_info_delete = {
@@ -21,15 +20,14 @@ export type T_user_info_delete_return = {
 };
 
 const user_info_delete = async (req: NextApiRequest, res: NextApiResponse) => {
-  if (req.method === "POST") {
-
+  if (req.method === 'POST') {
     // await runMiddleware(req, res);
 
     const { user_id }: T_user_info_delete = req.body;
 
     try {
       const data = await db(
-        "DELETE FROM `user_info` WHERE `user_id`=?",
+        'DELETE FROM `user_info` WHERE `user_id`=?',
         user_id
       );
 
@@ -39,37 +37,33 @@ const user_info_delete = async (req: NextApiRequest, res: NextApiResponse) => {
       // );
 
       const data3 = await db(
-        "DELETE FROM `articles` WHERE `user_id`=?",
+        'DELETE FROM `articles` WHERE `user_id`=?',
         user_id
       );
 
       const data4 = await db(
-        "DELETE FROM `footer_items` WHERE `user_id`=?",
+        'DELETE FROM `footer_items` WHERE `user_id`=?',
         user_id
       );
 
-      const data5 = await db(
-        "DELETE FROM `tags` WHERE `user_id`=?",
-        user_id
-      );
+      const data5 = await db('DELETE FROM `tags` WHERE `user_id`=?', user_id);
 
       const data6 = await db(
-        "DELETE FROM `instagram_accounts` WHERE `user_id`=?",
+        'DELETE FROM `instagram_accounts` WHERE `user_id`=?',
         user_id
       );
 
       const data7 = await db(
-        "DELETE FROM `info_bar` WHERE `user_id`=?",
+        'DELETE FROM `info_bar` WHERE `user_id`=?',
         user_id
       );
-      
+
       const returnData: T_user_info_delete_return = {
         rawData: data,
       };
       res.status(200).json(returnData);
-
     } catch (err) {
-      console.log("/user_info/delete/のエラーは " + JSON.stringify(err));
+      console.log('/user_info/delete/のエラーは ' + JSON.stringify(err));
       return res.status(500).json({ err: true, data: err });
     }
   }
@@ -81,9 +75,9 @@ export const config = {
   api: {
     externalResolver: true,
     bodyParser: {
-      sizeLimit: "50mb",
+      sizeLimit: '50mb',
     },
   },
 };
 
-export default user_info_delete
+export default user_info_delete;
